Adicionado verificação de User para evitar que demais usuários atrapalhem.

+++ src/main/java/com/br/projetoyaskara/service/LotesIngressosService.java
@@ -1,17 +1,20 @@
 package com.br.projetoyaskara.service;
 
 import com.br.projetoyaskara.dto.LotesIngressoDTO;
-import com.br.projetoyaskara.exception.BadRequestException;
 import com.br.projetoyaskara.exception.ResourceNotFoundException;
 import com.br.projetoyaskara.mapper.LotesIngressoMapper;
 import com.br.projetoyaskara.model.LotesIngresso;
+import com.br.projetoyaskara.model.Organizacao;
 import com.br.projetoyaskara.repository.EventosRepository;
 import com.br.projetoyaskara.repository.LotesIngressosRepository;
+import com.br.projetoyaskara.repository.UserRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.UUID;
 
 @Service
 public class LotesIngressosService {
@@ -19,40 +22,61 @@ public class LotesIngressosService {
     private final LotesIngressoMapper lotesIngressoMapper;
     private final LotesIngressosRepository lotesIngressosRepository;
     private final EventosRepository eventosRepository;
+    private final UserRepository userRepository;
 
     public LotesIngressosService(LotesIngressoMapper lotesIngressoMapper,
                                  LotesIngressosRepository lotesIngressosRepository,
-                                 EventosRepository eventosRepository) {
+                                 EventosRepository eventosRepository, UserRepository userRepository) {
         this.lotesIngressoMapper = lotesIngressoMapper;
         this.lotesIngressosRepository = lotesIngressosRepository;
         this.eventosRepository = eventosRepository;
+        this.userRepository = userRepository;
     }
 
 
     private LotesIngresso findLotesIngressoOrThrow(Long id) {
-        return lotesIngressosRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Ingresso não achado"));
+        return lotesIngressosRepository.
+                findById(id).orElseThrow(() -> new ResourceNotFoundException("Ingresso não achado"));
     }
 
     private void findEventoOrThrow(Long id) {
-        eventosRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Evento não encontrado"));
+        eventosRepository
+                .findById(id).orElseThrow(() -> new ResourceNotFoundException("Evento não encontrado"));
     }
 
-    public ResponseEntity<LotesIngressoDTO> cadastrarIngresso(LotesIngressoDTO lotesIngressoDTO) {
+    public ResponseEntity<LotesIngressoDTO> cadastrarIngresso(Authentication authentication,
+                                                              LotesIngressoDTO lotesIngressoDTO) {
 
-            findEventoOrThrow(lotesIngressoDTO.getIdEvento());
+        Organizacao organizacao = eventosRepository
+                .findByOrganizationIdByEventId(lotesIngressoDTO.getIdEvento());
 
-            if (lotesIngressoDTO.getDataInicio().isAfter(lotesIngressoDTO.getDataFim())) {
-                throw new BadRequestException("A data de início do lote não pode ser posterior à data de fim.");
-            }
+        UUID idUser = userRepository.findIdByEmail(authentication.getName());
 
-            LotesIngresso lotesIngresso = lotesIngressoMapper.toEntity(lotesIngressoDTO);
-            LotesIngresso ingressoSalvo = lotesIngressosRepository.save(lotesIngresso);
-            return ResponseEntity.status(HttpStatus.CREATED).body(lotesIngressoMapper.toDto(ingressoSalvo));
 
+        if (!eventosRepository.existsById(lotesIngressoDTO.getIdEvento())
+                || !organizacao.getProprietario().getId().equals(idUser))  {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        LotesIngresso lotesIngresso = lotesIngressoMapper.toEntity(lotesIngressoDTO);
+        LotesIngresso ingressoSalvo = lotesIngressosRepository.save(lotesIngresso);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(lotesIngressoMapper.toDto(ingressoSalvo));
     }
 
-    public ResponseEntity<LotesIngressoDTO> atualizarIngresso(LotesIngressoDTO lotesIngressoDTO) {
+    public ResponseEntity<LotesIngressoDTO> atualizarIngresso(Authentication authentication,
+                                                              LotesIngressoDTO lotesIngressoDTO) {
             LotesIngresso lotesIngressoExistente = findLotesIngressoOrThrow(lotesIngressoDTO.getId());
+
+            Organizacao organizacao = eventosRepository
+                    .findByOrganizationIdByEventId(lotesIngressoDTO.getIdEvento());
+
+            UUID idUser = userRepository.findIdByEmail(authentication.getName());
+
+            if (!eventosRepository.existsById(lotesIngressoDTO.getIdEvento())
+                || !organizacao.getProprietario().getId().equals(idUser))  {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
 
             lotesIngressoExistente.setName(lotesIngressoDTO.getName());
             lotesIngressoExistente.setValor(lotesIngressoDTO.getValor());
@@ -65,8 +89,18 @@ public class LotesIngressosService {
 
     }
 
-    public ResponseEntity<Void> deletarIngresso(Long id) {
-            LotesIngresso lotesIngresso = findLotesIngressoOrThrow(id);
+    public ResponseEntity<Void> deletarIngresso(Authentication authentication, Long id) {
+        LotesIngresso lotesIngresso = findLotesIngressoOrThrow(id);
+
+        Organizacao organizacao = eventosRepository
+                .findByOrganizationIdByEventId(lotesIngresso.getEvento().getId());
+
+        UUID idUser = userRepository.findIdByEmail(authentication.getName());
+
+            if (!eventosRepository.existsById(lotesIngresso.getEvento().getId())
+                || !organizacao.getProprietario().getId().equals(idUser))  {
+                 return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
             lotesIngressosRepository.delete(lotesIngresso);
             return ResponseEntity.ok().build();
     }
@@ -76,10 +110,8 @@ public class LotesIngressosService {
         if (lotesIngressos.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         }
-
         return ResponseEntity.status(HttpStatus.OK).body(lotesIngressoMapper.toDto(lotesIngressos));
     }
-
 
     public ResponseEntity<LotesIngressoDTO> buscarIngressoPorId(Long id) {
         LotesIngresso lotesIngresso = findLotesIngressoOrThrow(id);
@@ -92,19 +124,13 @@ public class LotesIngressosService {
         return ResponseEntity.status(HttpStatus.OK).body(lotesIngressoMapper.toDto(lotesIngresso));
     }
 
-    public ResponseEntity<List<LotesIngressoDTO>> buscarIngressosPorFaixaDePreco(int faixaMenorPreco, int faixaMaiorPreco) {
+    public ResponseEntity<List<LotesIngressoDTO>> buscarIngressosPorFaixaDePreco(int faixaMenorPreco,
+                                                                                 int faixaMaiorPreco) {
         List<LotesIngresso> lotesIngressos = lotesIngressosRepository
-                .findLotesIngressosPorFaixaDePreco(faixaMenorPreco, faixaMaiorPreco);
+                .findLotesIngressosPorFaixaDePreco(faixaMenorPreco,
+                                                    faixaMaiorPreco);
 
         return ResponseEntity.status(HttpStatus.OK).body(lotesIngressoMapper.toDto(lotesIngressos));
-    }
-
-    public ResponseEntity<?> aumentarVendaEm1(long id){
-        LotesIngresso lotesIngresso = findLotesIngressoOrThrow(id);
-
-        lotesIngresso.setTotalVendas(lotesIngresso.getTotalVendas() + 1);
-        lotesIngressosRepository.save(lotesIngresso);
-        return ResponseEntity.ok().body(lotesIngressoMapper.toDto(lotesIngresso));
     }
 
 }
